requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1671589829431/work
certifi==2023.5.7
cffi @ file:///D:/bld/cffi_1671179506518/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
click @ file:///D:/bld/click_1666798499870/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///D:/bld/cryptography-split_1681508778756/work
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1683130392307/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1673812078720/work
greenlet @ file:///D:/bld/greenlet_1684083063122/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
lxml @ file:///D:/bld/lxml_1671013694264/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydicom @ file:///home/conda/feedstock_root/build_artifacts/pydicom_1668525185822/work
pymongo @ file:///D:/bld/pymongo_1668739563954/work
pyodbc @ file:///D:/bld/pyodbc_1682444446063/work
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1677227146928/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy @ file:///D:/bld/sqlalchemy_1684569123315/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1683161174614/work
webdavclient3 @ file:///home/conda/feedstock_root/build_artifacts/webdavclient3_1616573459906/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wincertstore==0.2
